move OdooWebsiteEditor to subfolder
